+++ common/src/main/java/rzk/wirelessredstone/config/AttachmentMode.java
@@ -5,12 +5,20 @@ import rzk.wirelessredstone.misc.TranslationKeys;
 
 public enum AttachmentMode
 {
-	ATTACHED(Component.translatable(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ATTACHED)),
-	ADJACENT(Component.translatable(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ADJACENT));
+	ATTACHED(
+		Component.translatable(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ATTACHED),
+		Component.translatable(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ATTACHED_DESCRIPTION)
+	),
+	ADJACENT(
+		Component.translatable(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ADJACENT),
+		Component.translatable(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ADJACENT_DESCRIPTION)
+	);
 
-	public final Component component;
+	public final Component name;
+	public final Component description;
 
-	AttachmentMode(Component component) {
-		this.component = component;
+	AttachmentMode(Component name, Component description) {
+		this.name = name;
+		this.description = description;
 	}
 }