+++ datagen/src/main/java/rzk/wirelessredstone/datagen/DefaultLanguageGenerator.java
@@ -38,15 +38,18 @@ public class DefaultLanguageGenerator extends FabricLanguageProvider
 		translations.add(TranslationKeys.GUI_CONFIG_TITLE, "Wireless Redstone Config");
 		translations.add(TranslationKeys.GUI_CONFIG_CATEGORY_GENERAL, "General");
 		translations.add(TranslationKeys.GUI_CONFIG_CATEGORY_CLIENT, "Client");
-		translations.add(TranslationKeys.GUI_CONFIG_SIGNAL_STRENGTH, "Signal strength");
-		translations.add(TranslationKeys.GUI_CONFIG_STRONG_POWER, "Strong power");
-		translations.add(TranslationKeys.GUI_CONFIG_DISPLAY_COLOR, "Display color");
-		translations.add(TranslationKeys.GUI_CONFIG_HIGHLIGHT_COLOR, "Highlight color");
-		translations.add(TranslationKeys.GUI_CONFIG_HIGHLIGHT_TIME, "Highlight time");
+		translations.add(TranslationKeys.GUI_CONFIG_SIGNAL_STRENGTH, "Receiver Redstone signal strength");
+		translations.add(TranslationKeys.GUI_CONFIG_STRONG_POWER, "Receivers provide strong power");
+		translations.add(TranslationKeys.GUI_CONFIG_DISPLAY_COLOR, "Frequency display color");
+		translations.add(TranslationKeys.GUI_CONFIG_HIGHLIGHT_COLOR, "Sniffer highlight color");
+		translations.add(TranslationKeys.GUI_CONFIG_HIGHLIGHT_TIME, "Sniffer highlight time (seconds)");
 		translations.add(TranslationKeys.GUI_CONFIG_TARGET_COLOR, "Target highlight color");
+
 		translations.add(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE, "Attachment Mode");
 		translations.add(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ATTACHED, "Attached");
+		translations.add(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ATTACHED_DESCRIPTION, "Signal input and output only\nfrom and to the attached block itself");
 		translations.add(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ADJACENT, "Adjacent");
+		translations.add(TranslationKeys.GUI_CONFIG_ATTACHMENT_MODE_ADJACENT_DESCRIPTION, "Signal input and output from and to\nall adjacent blocks, except the one\nopposite to the attachment");
 
 		// Tooltips
 		translations.add(TranslationKeys.TOOLTIP_FREQUENCY, "Frequency: %s");
@@ -62,6 +65,8 @@ public class DefaultLanguageGenerator extends FabricLanguageProvider
 		translations.add(TranslationKeys.MESSAGE_TRANSMITTERS_ACTIVE, "Active transmitters on frequency %s: %s");
 		translations.add(TranslationKeys.MESSAGE_TELEPORT, "Teleport here");
 		translations.add(TranslationKeys.MESSAGE_NO_FREQUENCY, "No frequency set");
+		translations.add(TranslationKeys.MESSAGE_FREQUENCY_COPIED, "Frequency copied: %s");
+		translations.add(TranslationKeys.MESSAGE_FREQUENCY_SET, "Frequency set: %s");
 		translations.add(TranslationKeys.NETWORKING_FAILED, "Networking error: %s");
 		translations.add(TranslationKeys.MESSAGE_P2P_WRONG_TARGET, "Target is not a %s");
 		translations.add(TranslationKeys.MESSAGE_P2P_LINKED, "Linked to target at: %s");